+++ lite/core/optimizer/mir/fusion/__xpu__spatial_transformer_fuse_pass.cc
