+++ lite/kernels/xpu/__xpu__spatial_transformer_compute.h
