+++ lite/operators/CMakeLists.txt
@@ -251,7 +251,7 @@ add_operator(__xpu__squeeze_excitation_op extra SRCS __xpu__squeeze_excitation_o
 add_operator(__xpu__bigru_op extra SRCS __xpu__bigru_op.cc)
 add_operator(__xpu__dynamic_lstm_fuse_op extra SRCS __xpu__dynamic_lstm_fuse_op.cc)
 add_operator(__xpu__multi_softmax_op extra SRCS __xpu__multi_softmax_op.cc)
-add_operator(__xpu__spatial_transformer_op XPU extra SRCS __xpu__unet_spatial_transformer_op.cc)
+add_operator(__xpu__spatial_transformer_op XPU extra SRCS __xpu__spatial_transformer_op.cc)
 
 if(XPU_WITH_XFT)
     add_operator(fusion_decoding_op extra SRCS fusion_decoding_op.cc)
