+++ lite/operators/__xpu__spatial_transformer_op.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "lite/operators/__xpu__unet_spatial_transformer_op.h"
+#include "lite/operators/__xpu__spatial_transformer_op.h"
 #include <utility>
 #include <vector>
 #include "lite/core/op_registry.h"
